Split magic number and content checks.
