+++ openchrom/plugins/net.openchrom.csd.converter.supplier.gaml/plugin.xml
@@ -9,7 +9,8 @@
             filterName="GAML CSD Chromatogram (*.gaml)"
             id="net.openchrom.csd.converter.supplier.gaml.chromatogram"
             importConverter="net.openchrom.csd.converter.supplier.gaml.converter.ChromatogramImportConverter"
-            importMagicNumberMatcher="net.openchrom.csd.converter.supplier.gaml.converter.MagicNumberMatcher"
+            importMagicNumberMatcher="net.openchrom.xxd.converter.supplier.gaml.converter.MagicNumberMatcher"
+            importContentMatcher="net.openchrom.csd.converter.supplier.gaml.converter.FileContentMatcher"
             isExportable="false"
             isImportable="true">
       </ChromatogramSupplier>
