+++ openchrom/plugins/net.openchrom.xir.converter.supplier.gaml/plugin.xml
@@ -10,7 +10,8 @@
             filterName="IR Spectroscopy (*.gaml)"
             id="net.openchrom.xir.converter.supplier.gaml"
             importConverter="net.openchrom.xir.converter.supplier.gaml.converter.ScanImportConverter"
-            importMagicNumberMatcher="net.openchrom.xir.converter.supplier.gaml.converter.MagicNumberMatcher"
+            importMagicNumberMatcher="net.openchrom.xxd.converter.supplier.gaml.converter.MagicNumberMatcher"
+            importFileContentMatcher="net.openchrom.xir.converter.supplier.gaml.converter.FileContentMatcher"
             isExportable="false"
             isImportable="true">
       </ScanSupplier>
