+++ openchrom/plugins/net.openchrom.xxd.converter.supplier.gaml/src/net/openchrom/xxd/converter/supplier/gaml/converter/MagicNumberMatcher.java
@@ -1,0 +1,26 @@
+/*******************************************************************************
+ * Copyright (c) 2021, 2023 Lablicate GmbH.
+ *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ * Matthias Mailänder - initial API and implementation
+ *******************************************************************************/
+package net.openchrom.xxd.converter.supplier.gaml.converter;
+
+import java.io.File;
+
+import org.eclipse.chemclipse.converter.core.AbstractMagicNumberMatcher;
+import org.eclipse.chemclipse.converter.core.IMagicNumberMatcher;
+
+public class MagicNumberMatcher extends AbstractMagicNumberMatcher implements IMagicNumberMatcher {
+
+	@Override
+	public boolean checkFileFormat(File file) {
+
+		return checkFileExtension(file, ".gaml");
+	}
+}
